emit(kotlin,swift): extract hand-written runtime to runtime/{kotlin,swift}/
Kotlin and Swift were the only targets storing their hand-written
primitive runtime as Rust `const &str` raw-string blobs inside the
emitter, instead of as real source files under runtime/<lang>/ like
every other target (rust/ts/python/crystal/elixir/go all include_str!
their primitives from a runtime tree). That was path dependence, not a
language constraint — the blobs are verbatim source with zero
interpolation, so a `const X = r#"..."#` is a drop-in `const X =
include_str!(...)`.

Lift the 6 Kotlin + 8 Swift primitive blobs plus each target's
RH_TEST_SUPPORT (RoundhouseTestCase) into runtime/kotlin/*.kt and
runtime/swift/*.swift, byte-for-byte, and point the consts at them.
Build manifests (BUILD_GRADLE_KTS, PACKAGE_SWIFT*) deliberately stay
inline — that matches the cross-target convention for manifests
(rust2 CARGO_TOML_TEMPLATE, crystal SHARD_YML), which are assembled
with interpolation. App-specific test wiring (TestSetup.kt/.swift)
stays format!-built for the same reason.

Pure refactor, no behavior change: emitter consumes the same
&'static str by the same name; only the storage location moves. The
payoff is that the Kotlin/Swift runtime now gets language tooling
(ktlint/swiftformat, LSP, standalone compile) and cross-target runtime
edits land as plain file edits in all targets uniformly — the
prerequisite for emitting the assert_select Dom stub everywhere.

Verified byte-correct: kotlin + swift toolchain tests pass (emitted
projects compile and run — strict compilers would reject any escaping
drift); default cargo test green. Diff is 16 include_str! lines in,
1583 blob lines out.
