+++ src/emit/kotlin/primitives.rs
@@ -28,33 +28,7 @@ use crate::emit::EmittedFile;
 /// (`ActiveRecord::Base#fill_timestamps`). The emitter renders that chain
 /// as `Time.now().utc.iso8601` — a method call then two property reads —
 /// so `now()` returns a `TimeInstant` whose `utc`/`iso8601` are `val`s.
-const TIME_KT: &str = r#"// Hand-written roundhouse runtime primitive (no Ruby source).
-// Minimal shim for `Time.now.utc.iso8601`, used by
-// ActiveRecord::Base#fill_timestamps to stamp created_at/updated_at.
-
-package roundhouse
-
-import java.time.OffsetDateTime
-import java.time.ZoneOffset
-import java.time.format.DateTimeFormatter
-import java.time.temporal.ChronoUnit
-
-object Time {
-    fun now(): TimeInstant = TimeInstant(OffsetDateTime.now())
-}
-
-class TimeInstant(private val dt: OffsetDateTime) {
-    // `Time#utc` — the same instant at a UTC offset.
-    val utc: TimeInstant
-        get() = TimeInstant(dt.withOffsetSameInstant(ZoneOffset.UTC))
-
-    // `Time#iso8601` — seconds precision, `Z` for a zero offset (the `XXX`
-    // pattern renders UTC as `Z`, matching Ruby): `2026-06-07T17:30:00Z`.
-    val iso8601: String
-        get() = dt.truncatedTo(ChronoUnit.SECONDS)
-            .format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssXXX"))
-}
-"#;
+const TIME_KT: &str = include_str!("../../../runtime/kotlin/time.kt");
 
 /// The sqlite primitive layer the lowered model IR dispatches against
 /// (`Db.prepare` / `Db.step` / `Db.columnInt` / `Db.columnText` /
@@ -63,147 +37,7 @@ class TimeInstant(private val dt: OffsetDateTime) {
 /// renders `Ty::Int` literals with an `L` suffix (`Db.columnInt(stmt, 0L)`),
 /// so the column-index params are `Long` and shifted to JDBC's 1-based,
 /// `Int`-typed column index internally.
-const DB_KT: &str = r#"// Hand-written roundhouse runtime primitive (no Ruby source).
-// The sqlite primitive layer the lowered model IR dispatches against.
-// Mirrors `runtime/crystal/db.cr` and `runtime/typescript/db.ts`: an opaque
-// Long stmt id indexes a table of open JDBC ResultSets, and `step` /
-// `columnInt` / `columnText` read the cursor. The lowered `_adapter_*` model
-// emit calls exactly this surface.
-//
-// CONCURRENCY: Javalin/Jetty dispatches each request onto its own thread, and
-// the whole prepare → step → … → finalize sequence for a request runs
-// synchronously on that one thread. So all mutable cursor state is
-// THREAD-CONFINED via `ThreadLocal` — each thread lazily opens its own JDBC
-// connection and keeps its own statement table, counters, and last-insert
-// state. No locks, no shared maps: a JDBC `Connection` isn't safe for
-// concurrent use, and a shared `HashMap` corrupts under concurrent writes.
-// SQLite serves concurrent readers fine (a per-connection `busy_timeout`
-// rides out the brief writer lock). This is the per-thread analog of the
-// connection pool the crystal/go/rust runtimes size to wrk concurrency.
-//
-// xerial `sqlite-jdbc` is the locked driver. JDBC columns are 1-based, and
-// the emitter passes a zero-based `Long` index (Int literals carry an `L`
-// suffix), so the index is shifted and narrowed to `Int` here.
-
-package roundhouse
-
-import java.io.File
-import java.sql.Connection
-import java.sql.DriverManager
-import java.sql.PreparedStatement
-import java.sql.ResultSet
-
-object Db {
-    @Volatile private var path: String? = null
-
-    // One connection per thread, opened lazily against the configured path.
-    private val tlConn: ThreadLocal<Connection> = ThreadLocal.withInitial {
-        val p = path ?: error("Db not opened")
-        val c = DriverManager.getConnection("jdbc:sqlite:$p")
-        c.createStatement().use { it.execute("PRAGMA busy_timeout=5000") }
-        c
-    }
-    private val tlStatements: ThreadLocal<HashMap<Long, ResultSet>> =
-        ThreadLocal.withInitial { HashMap() }
-    private val tlOwners: ThreadLocal<HashMap<Long, PreparedStatement>> =
-        ThreadLocal.withInitial { HashMap() }
-    private val tlNextId: ThreadLocal<Long> = ThreadLocal.withInitial { 0L }
-    private val tlRowid: ThreadLocal<Long> = ThreadLocal.withInitial { 0L }
-    private val tlChanges: ThreadLocal<Long> = ThreadLocal.withInitial { 0L }
-
-    private fun conn(): Connection = tlConn.get()
-
-    fun openProductionDb(path: String) {
-        File(path).parentFile?.mkdirs()
-        this.path = path
-        // Open this (main) thread's connection eagerly to fail fast on a bad
-        // path; worker threads open theirs on first query.
-        tlConn.get()
-    }
-
-    // Per-test in-memory database: drop this thread's connection (a fresh
-    // jdbc:sqlite::memory: connection IS a fresh database) and replay the
-    // schema DDL. Tests dispatch synchronously on the JUnit thread, so the
-    // thread-local scope matches the test scope.
-    fun setupTestDb(schema: String) {
-        if (path != null) {
-            runCatching { tlConn.get().close() }
-        }
-        path = ":memory:"
-        tlStatements.remove()
-        tlOwners.remove()
-        tlConn.remove()
-        if (schema.isNotEmpty()) {
-            for (stmt in schema.split(";\n")) {
-                if (stmt.isNotBlank()) {
-                    exec(stmt)
-                }
-            }
-        }
-    }
-
-    // Run one-shot DDL/INSERT/UPDATE/DELETE; capture rowid + changes.
-    fun exec(sql: String) {
-        conn().createStatement().use { st ->
-            st.executeUpdate(sql)
-            conn().createStatement().use { c ->
-                c.executeQuery("SELECT last_insert_rowid(), changes()").use { rs ->
-                    if (rs.next()) {
-                        tlRowid.set(rs.getLong(1))
-                        tlChanges.set(rs.getLong(2))
-                    }
-                }
-            }
-        }
-    }
-
-    // Prepare a SELECT, returning an opaque integer handle.
-    fun prepare(sql: String): Long {
-        val ps = conn().prepareStatement(sql)
-        val rs = ps.executeQuery()
-        val id = tlNextId.get() + 1
-        tlNextId.set(id)
-        tlStatements.get()[id] = rs
-        tlOwners.get()[id] = ps
-        return id
-    }
-
-    // Advance the cursor; false (snapshot cleared) when exhausted.
-    fun step(stmtId: Long): Boolean {
-        val rs = tlStatements.get()[stmtId] ?: return false
-        return rs.next()
-    }
-
-    // Read an integer column at a zero-based index. NULL coerces to 0.
-    fun columnInt(stmtId: Long, i: Long): Long {
-        val rs = tlStatements.get()[stmtId]!!
-        val v = rs.getLong((i + 1).toInt())
-        return if (rs.wasNull()) 0L else v
-    }
-
-    // Read a text column at a zero-based index. NULL coerces to "".
-    fun columnText(stmtId: Long, i: Long): String {
-        val rs = tlStatements.get()[stmtId]!!
-        return rs.getString((i + 1).toInt()) ?: ""
-    }
-
-    // Release the ResultSet + statement. Idempotent.
-    fun finalize(stmtId: Long) {
-        tlStatements.get().remove(stmtId)?.close()
-        tlOwners.get().remove(stmtId)?.close()
-    }
-
-    fun lastInsertRowid(): Long = tlRowid.get()
-    fun changes(): Long = tlChanges.get()
-
-    // SQL-quote helpers the lowered adapter emit inlines.
-    fun escapeString(s: String): String = "'" + s.replace("'", "''") + "'"
-    fun escapeInt(n: Long): String = n.toString()
-    fun escapeBool(b: Boolean): String = if (b) "1" else "0"
-    // `IN (...)` list for a preload query: comma-joined integer ids.
-    fun escapeIntList(ids: MutableList<Long>): String = ids.joinToString(", ") { it.toString() }
-}
-"#;
+const DB_KT: &str = include_str!("../../../runtime/kotlin/db.kt");
 
 // The previous `ParamValue` sealed-union primitive was removed: the params
 // layer now holds untyped values as Kotlin's top type `Any?` (nested Hash →
@@ -222,213 +56,7 @@ object Db {
 /// formats the response (redirect, html-with-layout, or json). The routes
 /// table, controller factory map, and layout function are passed in by the
 /// emitted `Main.kt` (they're app-specific).
-const SERVER_KT: &str = r##"// Hand-written roundhouse runtime primitive (no Ruby source).
-// Javalin HTTP listener — parse request -> Router.match -> instantiate
-// controller -> process_action -> format response. Mirrors
-// runtime/crystal/server.cr.
-
-package roundhouse
-
-import io.javalin.Javalin
-import io.javalin.http.Context
-import io.javalin.http.Cookie
-import io.javalin.http.Handler
-
-object Server {
-    // Flash cookie name. Flash is cookie-backed and per-session (per
-    // browser), so parallel clients never share a flash slot — the
-    // "show exactly once" lifecycle lives in the transpiled Flash class
-    // (Flash.toPersisted keeps only what the action set), and dispatch is
-    // just the storage adapter. Mirrors go (server.go) / rust (http.rs).
-    private const val FLASH_COOKIE = "rh_flash"
-    // Compiled assets (tailwind.css, turbo.min.js, …) live under
-    // static/assets/ and are served at /assets/* by `serveAsset` (called
-    // from `dispatch`). Served in-handler rather than via Javalin's
-    // staticFiles because the greedy "/<path>" app route shadows Javalin's
-    // static handler. Absolute so it's independent of any cwd surprises.
-    private val assetsRoot: java.io.File = java.io.File("static/assets").absoluteFile
-
-    fun start(
-        dbPath: String,
-        port: Int,
-        routes: MutableList<Route>,
-        controllers: Map<String, () -> ActionControllerBase>,
-        layout: (String, String?, String?) -> String,
-    ) {
-        Db.openProductionDb(dbPath)
-        // Mount the Action Cable /cable WebSocket on Jetty's servlet context.
-        // Done at config time (not via app.ws) so the upgrade can negotiate
-        // the actioncable-v1-json subprotocol — see Cable / CableServlet.
-        val app = Javalin.create { config ->
-            config.jetty.modifyServletContextHandler { handler -> Cable.mount(handler) }
-        }
-        val handler = Handler { ctx -> dispatch(ctx, routes, controllers, layout) }
-        for (p in listOf("/", "/<path>")) {
-            app.get(p, handler)
-            app.post(p, handler)
-            app.put(p, handler)
-            app.patch(p, handler)
-            app.delete(p, handler)
-        }
-        app.start("127.0.0.1", port)
-        println("Roundhouse Kotlin server listening on http://127.0.0.1:$port")
-    }
-
-    // Serve a file from static/assets/, content-typed by extension.
-    // Path-traversal guarded (the canonical target must stay under the
-    // assets root). 404 when missing, so a fresh archive with no built
-    // assets still boots and the layout's asset links simply 404.
-    private fun serveAsset(ctx: Context, rel: String) {
-        val file = java.io.File(assetsRoot, rel).canonicalFile
-        if (!file.path.startsWith(assetsRoot.canonicalFile.path) || !file.isFile) {
-            ctx.status(404).result("Not Found")
-            return
-        }
-        val contentType = when (file.extension.lowercase()) {
-            "css" -> "text/css"
-            "js", "mjs" -> "application/javascript"
-            "json", "map" -> "application/json"
-            "svg" -> "image/svg+xml"
-            "png" -> "image/png"
-            else -> "application/octet-stream"
-        }
-        ctx.contentType(contentType).result(file.readBytes())
-    }
-
-    // Decode the rh_flash cookie into the String-keyed map the Flash
-    // constructor reloads from. Absent/empty -> empty map (first request
-    // in a session carries no flash). Only the closed notice/alert key
-    // set is surfaced. Values are percent-encoded (URLEncoder) so notice
-    // text survives the `key=value&…` structure + cookie-octet rules.
-    private fun readFlashCookie(ctx: Context): MutableMap<String, String> {
-        val out: MutableMap<String, String> = mutableMapOf()
-        val raw = ctx.cookie(FLASH_COOKIE) ?: return out
-        for (pair in raw.split("&")) {
-            val idx = pair.indexOf('=')
-            if (idx <= 0) continue
-            val k = pair.substring(0, idx)
-            if (k != "notice" && k != "alert") continue
-            val v = java.net.URLDecoder.decode(pair.substring(idx + 1), "UTF-8")
-            if (v.isNotEmpty()) out[k] = v
-        }
-        return out
-    }
-
-    // Persist the entries the action set (Flash.toPersisted already swept
-    // the show-once ones). Empty -> clear the cookie so a shown notice
-    // doesn't stick. HttpOnly + Path=/ to match go/rust.
-    private fun writeFlashCookie(ctx: Context, persisted: Map<String, String>) {
-        if (persisted.isEmpty()) {
-            ctx.removeCookie(FLASH_COOKIE, "/")
-            return
-        }
-        val parts = mutableListOf<String>()
-        for (k in listOf("notice", "alert")) {
-            persisted[k]?.let { parts.add("$k=" + java.net.URLEncoder.encode(it, "UTF-8")) }
-        }
-        ctx.cookie(Cookie(name = FLASH_COOKIE, value = parts.joinToString("&"), path = "/", isHttpOnly = true))
-    }
-
-    private fun dispatch(
-        ctx: Context,
-        routes: MutableList<Route>,
-        controllers: Map<String, () -> ActionControllerBase>,
-        layout: (String, String?, String?) -> String,
-    ) {
-        ViewHelpers.resetSlotsBang()
-
-        // Compiled assets (/assets/tailwind.css, …) — served before route
-        // dispatch so the greedy app router doesn't 404 them.
-        if (ctx.method().name == "GET" && ctx.path().startsWith("/assets/")) {
-            serveAsset(ctx, ctx.path().removePrefix("/assets/"))
-            return
-        }
-
-        // Rails' `_method` override (button_to delete/patch forms POST).
-        var method = ctx.method().name
-        if (method == "POST") {
-            ctx.formParam("_method")?.let { method = it.uppercase() }
-        }
-
-        // A `.json` extension selects the JSON variant.
-        var path = ctx.path()
-        var format = "html"
-        if (path.endsWith(".json")) {
-            format = "json"
-            path = path.substring(0, path.length - 5)
-        }
-
-        val match = Router.match(method, path, routes)
-        val factory = match?.let { controllers[it.controller] }
-        if (match == null || factory == null) {
-            ctx.status(404).result("Not Found")
-            return
-        }
-
-        val params: MutableMap<String, Any?> = mutableMapOf()
-        for ((k, v) in match.pathParams) {
-            params[k] = v
-        }
-        for ((k, vals) in ctx.queryParamMap()) {
-            vals.firstOrNull()?.let { setParam(params, k, it) }
-        }
-        for ((k, vals) in ctx.formParamMap()) {
-            vals.firstOrNull()?.let { setParam(params, k, it) }
-        }
-
-        val controller = factory()
-        controller.params = params
-        controller.requestFormat = format
-        controller.requestMethod = method
-        controller.requestPath = path
-        // Reload the flash carried from the previous request (the redirect
-        // that set `flash[:notice] = …`) so views render it; the
-        // constructor snapshots it as *_was so toPersisted can drop it
-        // after one display.
-        controller.flash = Flash(readFlashCookie(ctx))
-        controller.session = Session()
-        controller.processAction(match.action)
-
-        // Carry the flash the action SET into the next request (or clear it
-        // once shown). toPersisted keeps only entries changed this request,
-        // so a merely-displayed notice drops out — shows exactly once.
-        writeFlashCookie(ctx, controller.flash.toPersisted())
-
-        val code = controller.status.toInt()
-        val location = controller.location
-        if (location != null) {
-            ctx.status(code)
-            ctx.header("Location", location)
-        } else if (controller.requestFormat == "json") {
-            ctx.status(code).contentType("application/json").result(controller.body)
-        } else {
-            ctx.status(code)
-                .html(layout(controller.body, controller.flash.notice, controller.flash.alert))
-        }
-    }
-
-    // `article[title]=Foo` -> a nested `MutableMap<String, Any?>`; a bare
-    // key -> a scalar `String`. Untyped params are held as `Any?` (the
-    // Kotlin top type) so `<Resource>Params.from_raw`'s `is_a?(Hash)` /
-    // `is_a?(String)` narrowing — emitted as `is Map<*,*>` / `is String` —
-    // matches against real Map/String values rather than a wrapper that
-    // would fail every check.
-    private fun setParam(params: MutableMap<String, Any?>, key: String, value: String) {
-        val open = key.indexOf('[')
-        if (open >= 0 && key.endsWith("]")) {
-            val outer = key.substring(0, open)
-            val inner = key.substring(open + 1, key.length - 1)
-            val existing = params[outer]
-            @Suppress("UNCHECKED_CAST")
-            val dict = if (existing is MutableMap<*, *>) existing as MutableMap<String, Any?> else mutableMapOf()
-            dict[inner] = value
-            params[outer] = dict
-        } else {
-            params[key] = value
-        }
-    }
-}
-"##;
+const SERVER_KT: &str = include_str!("../../../runtime/kotlin/server.kt");
 
 /// The adapter contract `ActiveRecord::Base`'s class-level CRUD defaults
 /// (`_adapter_all` / `_adapter_find_by_id` / `where` / `find_by` / …)
@@ -447,26 +75,7 @@ object Server {
 /// per-model override are `where`/`find_by`, which real-blog never invokes
 /// and which therefore throw `UninitializedPropertyAccessException` if hit
 /// — the correct "this path is unsupported" behavior.
-const ADAPTER_INTERFACE_KT: &str = r#"// Hand-written roundhouse runtime primitive (no Ruby source).
-// The adapter contract ActiveRecord::Base's class-level CRUD defaults
-// type-check against. Surface mirrors active_record/base.rbs's
-// AdapterInterface. The legacy functional adapter path is dropped for
-// Kotlin — no implementation is provided and `ActiveRecord.adapter` is
-// never wired; all real CRUD is Db-direct via the Level-3 per-model
-// `_adapter_*` overrides. This interface only lets Base's (unreached)
-// defaults compile.
-
-package roundhouse
-
-interface AdapterInterface {
-    fun all(tableName: String): MutableList<MutableMap<String, Any?>>
-    fun find(tableName: String, id: Long): MutableMap<String, Any?>?
-    fun where(tableName: String, conditions: MutableMap<String, Any?>): MutableList<MutableMap<String, Any?>>
-    fun count(tableName: String): Long
-    fun exists(tableName: String, id: Long): Boolean
-    fun truncate(tableName: String)
-}
-"#;
+const ADAPTER_INTERFACE_KT: &str = include_str!("../../../runtime/kotlin/adapter_interface.kt");
 
 /// Turbo Streams broadcast sink — the object the model `after_*_commit`
 /// callbacks dispatch to (`Broadcasts.append`/`prepend`/`replace`/`remove`,
@@ -474,27 +83,7 @@ interface AdapterInterface {
 /// `stream`/`target`/`html`). Composes the `<turbo-stream>` fragment and
 /// hands it to the cable fan-out (`Cable.dispatch`). Mirrors the
 /// go2/rust2/crystal Broadcasts shim.
-const BROADCASTS_KT: &str = r#"// Hand-written roundhouse runtime primitive (no Ruby source).
-// Turbo Streams broadcast sink. The model after_*_commit callbacks pass a
-// {stream, target, html} bag; compose the <turbo-stream> wrapper and fan it
-// out to /cable subscribers via Cable. Mirrors go/rust/crystal's Broadcasts.
-
-package roundhouse
-
-object Broadcasts {
-    fun append(opts: MutableMap<String, Any?>) = record("append", opts)
-    fun prepend(opts: MutableMap<String, Any?>) = record("prepend", opts)
-    fun replace(opts: MutableMap<String, Any?>) = record("replace", opts)
-    fun remove(opts: MutableMap<String, Any?>) = record("remove", opts)
-
-    private fun record(action: String, opts: MutableMap<String, Any?>) {
-        val stream = opts["stream"] as? String ?: return
-        val target = opts["target"] as? String ?: ""
-        val html = opts["html"] as? String ?: ""
-        Cable.dispatch(stream, Cable.turboStreamHtml(action, target, html))
-    }
-}
-"#;
+const BROADCASTS_KT: &str = include_str!("../../../runtime/kotlin/broadcasts.kt");
 
 /// Action Cable WebSocket + Turbo Streams broadcaster — the per-target
 /// transport primitive (cf. `runtime/go/v2/cable.go`,
@@ -507,166 +96,7 @@ object Broadcasts {
 /// ActionCable requires (javalin#957) — and the client closes the socket
 /// without it. The servlet's creator sets the accepted subprotocol DURING the
 /// upgrade. Server.kt mounts it via `config.jetty.modifyServletContextHandler`.
-const CABLE_KT: &str = r##"// Hand-written roundhouse runtime primitive (no Ruby source).
-// Action Cable WebSocket + Turbo Streams broadcaster (actioncable-v1-json).
-// Raw Jetty 11 WebSocket servlet so the upgrade can negotiate the
-// `actioncable-v1-json` subprotocol — Javalin's app.ws upgrades before its
-// onConnect runs, too late to set the header ActionCable requires. Mirrors
-// runtime/{go/v2,crystal,rust}/cable.
-
-package roundhouse
-
-import java.util.Base64
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.CopyOnWriteArrayList
-import java.util.concurrent.Executors
-import java.util.concurrent.TimeUnit
-import org.eclipse.jetty.servlet.ServletContextHandler
-import org.eclipse.jetty.servlet.ServletHolder
-import org.eclipse.jetty.websocket.api.Session
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketClose
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketConnect
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage
-import org.eclipse.jetty.websocket.api.annotations.WebSocket
-import org.eclipse.jetty.websocket.server.JettyWebSocketServlet
-import org.eclipse.jetty.websocket.server.JettyWebSocketServletFactory
-import org.eclipse.jetty.websocket.server.config.JettyWebSocketServletContainerInitializer
-import org.json.JSONObject
-import org.json.JSONTokener
-
-object Cable {
-    private data class Sub(val session: Session, val identifier: String)
-
-    // channel name -> live subscriptions. The identifier (the raw subscribe
-    // frame's `identifier` string) is echoed on every broadcast so Turbo
-    // routes the frame to the right <turbo-cable-stream-source>.
-    private val subscribers = ConcurrentHashMap<String, CopyOnWriteArrayList<Sub>>()
-    private val sessions = CopyOnWriteArrayList<Session>()
-    private val pinger = Executors.newSingleThreadScheduledExecutor { r ->
-        Thread(r, "cable-ping").apply { isDaemon = true }
-    }
-    @Volatile private var pingerStarted = false
-
-    // Mount the /cable servlet on Javalin's Jetty context. An exact servlet
-    // mapping takes precedence over Javalin's greedy "/<path>" route.
-    fun mount(handler: ServletContextHandler) {
-        JettyWebSocketServletContainerInitializer.configure(handler, null)
-        handler.addServlet(ServletHolder(CableServlet()), "/cable")
-        synchronized(this) {
-            if (!pingerStarted) {
-                pingerStarted = true
-                // ActionCable clients treat a ping gap (~6s) as a dead
-                // connection and reconnect, so heartbeat every 3s.
-                pinger.scheduleAtFixedRate({ pingAll() }, 3, 3, TimeUnit.SECONDS)
-            }
-        }
-    }
-
-    fun onConnect(session: Session) {
-        sessions.add(session)
-        safeSend(session, JSONObject().put("type", "welcome").toString())
-    }
-
-    fun onMessage(session: Session, message: String) {
-        val frame = try { JSONObject(message) } catch (e: Exception) { return }
-        if (frame.optString("command") != "subscribe") return
-        val identifier = frame.optString("identifier")
-        if (identifier.isEmpty()) return
-        val channel = decodeChannel(identifier) ?: return
-        subscribers.computeIfAbsent(channel) { CopyOnWriteArrayList() }.add(Sub(session, identifier))
-        safeSend(
-            session,
-            JSONObject().put("type", "confirm_subscription").put("identifier", identifier).toString(),
-        )
-    }
-
-    fun onClose(session: Session) {
-        sessions.remove(session)
-        for ((channel, subs) in subscribers) {
-            subs.removeAll { it.session === session }
-            if (subs.isEmpty()) subscribers.remove(channel, subs)
-        }
-    }
-
-    // Fan `html` out to every subscriber of `channel`, wrapped in the Action
-    // Cable message envelope Turbo expects. Called from Broadcasts on each
-    // model after-commit hook.
-    fun dispatch(channel: String, html: String) {
-        val subs = subscribers[channel] ?: return
-        for (sub in subs) {
-            val msg = JSONObject()
-                .put("type", "message")
-                .put("identifier", sub.identifier)
-                .put("message", html)
-                .toString()
-            safeSend(sub.session, msg)
-        }
-    }
-
-    fun turboStreamHtml(action: String, target: String, content: String): String =
-        if (content.isEmpty())
-            "<turbo-stream action=\"$action\" target=\"$target\"></turbo-stream>"
-        else
-            "<turbo-stream action=\"$action\" target=\"$target\"><template>$content</template></turbo-stream>"
-
-    private fun pingAll() {
-        val now = System.currentTimeMillis() / 1000
-        val frame = JSONObject().put("type", "ping").put("message", now).toString()
-        for (session in sessions) safeSend(session, frame)
-    }
-
-    // Jetty's blocking sendString throws on concurrent sends to one socket
-    // (a broadcast fiber racing the ping thread, or two creates fanning out
-    // to a shared subscriber), so serialize per session.
-    private fun safeSend(session: Session, msg: String) {
-        if (!session.isOpen) return
-        try {
-            synchronized(session) { session.remote.sendString(msg) }
-        } catch (e: Exception) {
-            // socket closed between the snapshot and the write — onClose cleans up
-        }
-    }
-
-    // Recover the channel name from Turbo's signed_stream_name. The identifier
-    // is `{"channel":"Turbo::StreamsChannel","signed_stream_name":"<b64>--<digest>"}`;
-    // the base64 prefix decodes to a JSON-encoded stream name (the same string
-    // a broadcast's `stream` carries). Returns null on malformed input.
-    private fun decodeChannel(identifier: String): String? = try {
-        val signed = JSONObject(identifier).optString("signed_stream_name")
-        val b64 = signed.substringBefore("--")
-        val decoded = String(Base64.getDecoder().decode(b64))
-        JSONTokener(decoded).nextValue() as? String
-    } catch (e: Exception) {
-        null
-    }
-}
-
-// The Jetty WebSocket servlet for /cable. Its creator sets the accepted
-// subprotocol during the upgrade handshake — the one thing Javalin's ws API
-// can't do.
-class CableServlet : JettyWebSocketServlet() {
-    override fun configure(factory: JettyWebSocketServletFactory) {
-        factory.setCreator { req, resp ->
-            if (req.subProtocols.contains("actioncable-v1-json")) {
-                resp.acceptedSubProtocol = "actioncable-v1-json"
-            }
-            CableEndpoint()
-        }
-    }
-}
-
-@WebSocket
-class CableEndpoint {
-    @OnWebSocketConnect
-    fun onConnect(session: Session) = Cable.onConnect(session)
-
-    @OnWebSocketMessage
-    fun onMessage(session: Session, message: String) = Cable.onMessage(session, message)
-
-    @OnWebSocketClose
-    fun onClose(session: Session, statusCode: Int, reason: String?) = Cable.onClose(session)
-}
-"##;
+const CABLE_KT: &str = include_str!("../../../runtime/kotlin/cable.kt");
 
 /// The hand-written runtime primitives, emitted under `src/main/kotlin/`.
 pub fn primitives() -> Vec<EmittedFile> {
